requirements.txt
affine==2.3.0
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
contextily==1.1.0
cycler==0.10.0
fastapi==0.67.0
Fiona==1.8.20
geographiclib==1.52
geopandas==0.9.0
geopy==2.2.0
h11==0.12.0
httptools==0.2.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
kml2geojson==4.0.2
matplotlib==3.4.2
mercantile==1.2.1
munch==2.5.0
numpy==1.21.1
pandas==1.3.0
Pillow==8.3.1
pydantic==1.8.2
pyparsing==2.4.7
pyproj==3.1.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
rasterio==1.2.6
requests==2.26.0
Shapely==1.7.1
six==1.16.0
snuggs==1.4.7
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1